ADC to PSI
